refactor: UserTicket 엔티티 제거
