+++ app/src/androidTest/java/woowacourse/movie/ui/complete/MovieReservationCompleteActivityTest.kt
@@ -15,8 +15,8 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.R
 import woowacourse.movie.model.data.UserTicketsImpl
-import woowacourse.movie.model.db.UserTicket
 import woowacourse.movie.model.movie.ReservationDetail
+import woowacourse.movie.model.movie.UserTicket
 import woowacourse.movie.ui.selection.MovieSeatSelectionKey
 import java.time.LocalDateTime
