+++ app/src/main/java/woowacourse/movie/model/movie/UserTicket.kt
@@ -1,0 +1,11 @@
+package woowacourse.movie.model.movie
+
+import java.time.LocalDateTime
+
+data class UserTicket(
+    val title: String,
+    val theater: String,
+    val screeningStartDateTime: LocalDateTime,
+    val reservationDetail: ReservationDetail,
+    val id: Long = 0,
+)
