+++ app/src/main/java/woowacourse/movie/ui/complete/MovieReservationCompleteContract.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie.ui.complete
 
-import woowacourse.movie.model.db.UserTicket
+import woowacourse.movie.model.movie.UserTicket
 import woowacourse.movie.ui.HandleError
 
 interface MovieReservationCompleteContract {
