+++ app/src/main/res/layout-land/activity_movie_reservation_complete.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="userTicket"
-            type="woowacourse.movie.model.db.UserTicket" />
+            type="woowacourse.movie.model.movie.UserTicket" />
     </data>
 
     <LinearLayout
